pyproject.toml
[project]
name = "resume-career-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-search-results>=2.4.2",
    "jinja2>=3.1.6",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.8",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "serpapi>=0.1.5",
    "uvicorn>=0.35.0",
]
